evaluations/requirements.txt
# Offline Evaluation Dependencies
# These are the packages needed to run the offline evaluator

# Core dependencies
elasticsearch>=8.0.0
python-dotenv>=1.0.0

# Azure AI Evaluation SDK
azure-ai-evaluation>=1.0.0

# OpenTelemetry for tracing
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-elasticsearch>=0.40b0

# HTTP client for Azure services
httpx>=0.24.0
requests>=2.31.0

# JSON handling
jsonschema>=4.17.0

# Logging and utilities
tqdm>=4.65.0
